+++ pkg/frontend/querymiddleware/results_cache_test.go
@@ -7,7 +7,6 @@ package querymiddleware
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"testing"
 	"time"
@@ -40,8 +39,9 @@ func TestResultsCacheConfig_Validate(t *testing.T) {
 			cfg: ResultsCacheConfig{
 				BackendConfig: cache.BackendConfig{
 					Backend: cache.BackendMemcached,
-					Memcached: mimir_tsdb.MemcachedConfig{
-						Addresses: "localhost",
+					Memcached: mimir_tsdb.MemcachedClientConfig{
+						Addresses:           []string{"localhost"},
+						MaxAsyncConcurrency: 1,
 					},
 				},
 			},
@@ -50,12 +50,12 @@ func TestResultsCacheConfig_Validate(t *testing.T) {
 			cfg: ResultsCacheConfig{
 				BackendConfig: cache.BackendConfig{
 					Backend: cache.BackendMemcached,
-					Memcached: mimir_tsdb.MemcachedConfig{
-						Addresses: "",
+					Memcached: mimir_tsdb.MemcachedClientConfig{
+						Addresses: nil,
 					},
 				},
 			},
-			expected: errors.New("query-frontend results cache: no memcached addresses configured"),
+			expected: cache.ErrNoMemcachedAddresses,
 		},
 		"should fail with unsupported backend": {
 			cfg: ResultsCacheConfig{
@@ -63,14 +63,14 @@ func TestResultsCacheConfig_Validate(t *testing.T) {
 					Backend: "unsupported",
 				},
 			},
-			expected: errUnsupportedResultsCacheBackend("unsupported"),
+			expected: errUnsupportedBackend,
 		},
 	}
 
 	for testName, testData := range tests {
 		t.Run(testName, func(t *testing.T) {
 			if testData.expected != nil {
-				assert.EqualError(t, testData.cfg.Validate(), testData.expected.Error())
+				assert.ErrorIs(t, testData.cfg.Validate(), testData.expected)
 			} else {
 				assert.NoError(t, testData.cfg.Validate())
 			}