+++ tools/doc-generator/parse/root_blocks.go
@@ -152,8 +152,13 @@ var (
 		},
 		{
 			Name:       "memcached",
-			StructType: reflect.TypeOf(cache.MemcachedConfig{}),
+			StructType: reflect.TypeOf(cache.MemcachedClientConfig{}),
 			Desc:       "The memcached block configures the Memcached-based caching backend.",
+		},
+		{
+			Name:       "redis",
+			StructType: reflect.TypeOf(cache.RedisClientConfig{}),
+			Desc:       "The redis block configures the Redis-based caching backend.",
 		},
 		{
 			Name:       "s3_storage_backend",